+++ scripts/run.sh
@@ -152,13 +152,6 @@ while true; do
         patch "$2"
         break
         ;;
-    terraform)
-        # shellcheck source=scripts/commands/downloader.sh
-        . "${SCRIPT_DIR}/commands/terraform.sh"
-
-        terraform "$2"
-        break
-        ;;
     --help | -h | help)
         # shellcheck source=scripts/commands/help.sh
         . "${SCRIPT_DIR}/commands/help.sh"